+++ packages/core/src/index.ts
@@ -49,7 +49,15 @@ import Settings from './Settings';
 import * as volumeLoader from './loaders/volumeLoader';
 import * as imageLoader from './loaders/imageLoader';
 import * as geometryLoader from './loaders/geometryLoader';
-import * as Types from './types';
+import ProgressiveRetrieveImages from './loaders/ProgressiveRetrieveImages';
+import type * as Types from './types';
+import {
+  IRetrieveConfiguration,
+  IImagesLoader,
+  RetrieveOptions,
+  RetrieveStage,
+  ImageLoadListener,
+} from './types';
 import * as utilities from './utilities';
 import { registerImageLoader } from './loaders/imageLoader'; // since it is used by CSWIL right now
 
@@ -60,7 +68,15 @@ import {
   addVolumesToViewports,
 } from './RenderingEngine/helpers';
 
-export type { Types };
+// Add new types here so that they can be imported singly as required.
+export type {
+  Types,
+  IRetrieveConfiguration,
+  RetrieveOptions,
+  RetrieveStage,
+  ImageLoadListener,
+  IImagesLoader,
+};
 
 export {
   // init
@@ -124,4 +140,5 @@ export {
   resetUseSharedArrayBuffer,
   // Geometry Loader
   geometryLoader,
+  ProgressiveRetrieveImages,
 };